fixes footer visiblity
