+++ src/App.jsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react'
 import './App.css';
 import Footer from './Components/Footer';
 import { Switch, Route } from 'react-router-dom';
@@ -8,18 +8,27 @@ import Portfolio from './Components/Portfolio';
 import Contact from './Components/Contact';
 
 const App = () => {
-  
+  const [matrix, setMatrix] = useState('0vh');
+  const [blank, setBlank] = useState('visible');
 
+const pullBack = () => {
+  setBlank('hidden');
+  setMatrix('cover');
+  setTimeout(function(){
+    setMatrix('0vh');
+    setBlank('visible');
+},5000); 
+}
 
   return (
     <>
     <Switch>
-    <Route exact path='/' component={Landing}></Route> 
+    <Route exact path='/' render={() => ( <Landing blank={blank} matrix={matrix} pullBack={pullBack}/>)}></Route> 
     <Route exact path='/about' component={About}></Route>
     <Route exact path='/portfolio' component={Portfolio}></Route>
     <Route exact path='/contact' component={Contact}></Route>
     </Switch>
-    <Footer />
+    <Footer blank={blank} pullBack={pullBack}/>
    </>
   );
 }
