+++ src/Components/Footer.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { NavLink, Link } from "react-router-dom";
 
-const Footer = ({blank, setBlank }) => {
+const Footer = ({blank}) => {
   // const [blank, setBlank] = useState('visible');
   const name = "Steve Watson";
   const currentDate = new Date();
